+++ app/Services/CouncilSessionService.php
@@ -25,16 +25,34 @@ class CouncilSessionService
 
     public function createSession(array $details)
     {
-        return $this->sessionRepository->createSession($details);
+        $user = auth()->user();
+        if($user->role->name == 'VicePresident' || ($user->role->name == 'Head' && $user->council_id == $details['council_id'])){
+            return $this->sessionRepository->createSession($details);
+        }
+        return response()->json([
+            'message' => 'You are not authorized to create this session',
+        ], 403);
     }
 
     public function updateSession($id, array $details)
     {
-        return $this->sessionRepository->updateSession($id, $details);
+        $user = auth()->user();
+        if($user->role->name == 'VicePresident' || ($user->role->name == 'Head' && $user->council_id == $details['council_id'])){
+            return $this->sessionRepository->updateSession($id, $details);
+        }
+        return response()->json([
+            'message' => 'You are not authorized to update this session',
+        ], 403);
     }
 
     public function deleteSession($id)
     {
-        return $this->sessionRepository->deleteSession($id);
+        $user = auth()->user();
+        if($user->role->name == 'VicePresident' || (($user->role->name == 'Head'|| $user->role->name == 'Instructor') && $user->council_id == $id)){
+            return $this->sessionRepository->deleteSession($id);
+        }
+        return response()->json([
+            'message' => 'You are not authorized to delete this session',
+        ], 403);
     }
 }
